Split migration and persistence tests
